crud works

+++ server/api/logs.js
@@ -1,12 +1,13 @@
+/* eslint-disable no-shadow */
 /* eslint-disable no-underscore-dangle */
 /* eslint-disable no-multiple-empty-lines */
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-console */
 const { Router } = require('express');
+const mongoose = require('mongoose');
 const MongoStore = require('rate-limit-mongo');
 
 const LogEntry = require('../models/Logentry');
-// const User = require('../models/User');
 
 
 const router = Router();
@@ -35,26 +36,59 @@ router.post('/', async (req, res, next) => {
   }
 });
 
-// UpdateLog
-router.get('/logentry/:id/edit', (req, res, next) => {
+// Grab a specific log
+
+router.get('/logentries/:id', (req, res, next) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).json({ message: 'Specified id is not valid' });
+    return;
+  }
+
+  // we can use .populate() method to get the whole task objects
   LogEntry.findById(req.params.id)
-    .then((foundLogentry) => {
-      res.render('logentry/update-form', foundLogentry);
+    .populate('tasks')
+    // eslint-disable-next-line arrow-parens
+    .then(LogEntry => {
+      res.status(200).json(LogEntry);
     })
-    .catch((err) => {
-      // eslint-disable-next-line no-console
-      console.log(`error finding logentry by id due to ${err}`);
+    .catch((error) => {
+      res.json(error);
+    });
+});
+
+// UpdateLog
+router.put('/logentries/:id', (req, res, next) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).json({ message: 'Specified id is not valid' });
+    return;
+  }
+
+  LogEntry.findByIdAndUpdate(req.params.id, req.body)
+    .then(() => {
+      res.json({
+        message: `Log Entry with ${req.params.id} is updated successfully.`,
+      });
+    })
+    .catch((error) => {
+      res.json(error);
     });
 });
 
 // // Delete a log
-router.post('/Logentry/:id/delete', (req, res, next) => {
+router.delete('/logentries/:id', (req, res, next) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).json({ message: 'Specified id is not valid' });
+    return;
+  }
+
   LogEntry.findByIdAndRemove(req.params.id)
     .then(() => {
-      res.redirect('/Logentry');
+      res.json({
+        message: `Log entry with ${req.params.id} is removed successfully.`,
+      });
     })
-    .catch((err) => {
-      console.log(`err deleting drone due to ${err}`);
+    .catch((error) => {
+      res.json(error);
     });
 });
 
@@ -76,9 +110,5 @@ router.post('/Logentry/:id/delete', (req, res, next) => {
 //     user = await User.create(req.body);
 //   }
 // });
-
-
-
-
 
 module.exports = router;
